requirements.txt
sherpa-onnx
torch
torchaudio
numpy
scipy
soundfile
python-dotenv